Fixed release workflow
When uploading an asset to a release with the `gh` cli, you can specify
a display name. That way, you can upload your binary with the default
name while presenting it as something more descriptive, like adding the
architecture at the end of the file name. Then, when a user downloads
it, they receive the binary file without the architecture at the end.
The user does not need to rename the file to use it as intended.
However! Apparently trying to upload two assets with the same actual
names with two different display names still results in a collision.
Unintuitive behavior, but an easy fix.